backend/requirements/base.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter>=25.1
drf-spectacular>=0.28.0
psycopg[binary]>=3.2.7
python-decouple>=3.8
Pillow>=11.2.1
gunicorn>=23.0.0
redis>=6.0.0
celery>=5.6.0
django-redis==6.0.0
flower>=2.0.1
channels==4.3.2
channels-redis>=4.2.1
uvicorn>=0.34.0
openpyxl>=3.1.5
pytest>=8.3.5
pytest-django>=4.11.1
factory_boy>=3.3.3
black>=25.1.0
flake8>=7.2.0
isort>=6.0.1
